annotate models
https://guides.rubyonrails.org/v3.2/migrations.html
6.1 What are Schema Files for?
"Schema files are also useful if you want a quick look at what attributes an Active Record object has. This information is not in the model’s code and is frequently spread across several migrations, but the information is nicely summed up in the schema file. The annotate_models gem automatically adds and updates comments at the top of each model summarizing the schema if you desire that functionality."